* libawn/awn-applet.c: 	* libawn/awn-applet.h: 	* src/awn-panel-dbus.xml: 	* src/awn-panel.c: 	Added docklet_request method to AwnApplet.

+++ src/awn-panel.c
@@ -2444,7 +2444,7 @@ awn_panel_docklet_request (AwnPanel *panel,
                            DBusGMethodInvocation *context)
 {
   AwnPanelPrivate *priv = panel->priv;
-  guint32 window_id = 0;
+  gint64 window_id = 0;
 
   if (!priv->docklet)
   {
